Improve formatting of time on screen
Instead of just rounding, use string manipulation of the time double to print a time in the format 123.123456. Unlikely it's useful to have an integer part of over 1000 for this timer and there's not enough resolution on a 120Hz refresh screen to show more than 5 or 6 digits of the fractional part of the CPU time.